Create symbol table only at interpret time
Before we were adding variables in the symbol table while parsing, which
was... weird.  If we ever move from an interpreter to a compiler, this
would have been even more so.

With this massive commit, we move the symbol table manipulation to the
interpretation phase.  We also create symbols for the known types (int,
float, string for now).

While the Block struct was a cute way to handle scoping, it was a hack.
Since this is eminently a recursive problem, treat it properly by
creating / destroying a new nested symbol table every time we enter /
leave a scope.  So long Block.